refactor(examples): extract stacio and validate modules
Move STAC assembly, manifest loading, providers, assets, links, sidecars,
and the catalog builders into stacio.py, and the schema plus Portolan
conformance checks into validate.py. build.py is now an 89-line entrypoint,
only the PEP 723 header, argparse, path resolution, and main.

Repoint tests/check_compliance.py off "import build" onto the new stacio
and validate modules, and update tests/check_web_output.py to name convert.

Pure mechanical split, no behavior change. A full rebuild produces a
byte-identical examples/catalog tree and validation still passes.
